Food_Customer_Chatbot/backend/requirements.txt
emoji==2.14.1
fastapi==0.115.12
psycopg2-binary==2.9.9
pydantic==2.11.3
sentence_transformers==4.0.1
uvicorn==0.34.0
python-dotenv==1.1.0
